All double precision

+++ bark/benchmark/benchmark_runner.py
@@ -199,7 +199,7 @@ class BenchmarkRunner:
             self._append_to_scenario_history(scenario_history, world, scenario)
         self._reset_evaluators(world, scenario._eval_agent_ids)
         step_time = parameter_server["Simulation"]["StepTime", "", 0.2]
-        if not isinstance(step_time, float):
+        if not isinstance(step_time, double):
             step_time = 0.2
         terminal = False
         terminal_why = None
