+++ bark/models/behavior/behavior_rss/behavior_rss.hpp
@@ -63,7 +63,7 @@ class BehaviorRSSConformant : public BehaviorModel {
 
   virtual ~BehaviorRSSConformant() {}
 
-  Trajectory Plan(float min_planning_time, const ObservedWorld& observed_world);
+  Trajectory Plan(double min_planning_time, const ObservedWorld& observed_world);
 
   virtual std::shared_ptr<BehaviorModel> Clone() const;
 
@@ -105,7 +105,7 @@ class BehaviorRSSConformant : public BehaviorModel {
   std::shared_ptr<BehaviorSafety> behavior_safety_model_;
   std::shared_ptr<BaseEvaluator> rss_evaluator_;
   BehaviorRSSConformantStatus behavior_rss_status_;
-  float world_time_of_last_rss_violation_;
+  double world_time_of_last_rss_violation_;
   LaneCorridorPtr initial_lane_corr_;
   #ifdef RSS
   ::ad::rss::state::LongitudinalResponse lon_response_;
