+++ bark/models/dynamic/integration.hpp
@@ -20,7 +20,7 @@ namespace dynamic {
 inline State euler_int(const DynamicModel& model,
                 const State &x,
                 const Input &u,
-                float dt) {
+                double dt) {
   State new_x = x + dt * model.StateSpaceModel(x, u);
   new_x(StateDefinition::THETA_POSITION) =
       geometry::Norm0To2PI(new_x(StateDefinition::THETA_POSITION));
@@ -28,7 +28,7 @@ inline State euler_int(const DynamicModel& model,
 }
 
 inline State rk4(const DynamicModel& model, const State& x, const Input& u,
-                 float dt) {
+                 double dt) {
   State k0 = dt * model.StateSpaceModel(x, u);
   State k1 = dt * model.StateSpaceModel(x + k0 / 2, u);
   State k2 = dt * model.StateSpaceModel(x + k1 / 2, u);
