+++ bark/runtime/commons/parameters.py
@@ -240,19 +240,19 @@ class ParameterServer(Params):
 
     def GetCondensedParamList(self):
         def check_append(value):
-          if isinstance(value, float) or isinstance(value, int) \
+          if isinstance(value, double) or isinstance(value, int) \
               or isinstance(value,bool) or isinstance(value, str):
               return True
-          # list float
-          elif isinstance(value, list) and all(isinstance(el, float) for el in value):
+          # list double
+          elif isinstance(value, list) and all(isinstance(el, double) for el in value):
             return True
-          # list list float
+          # list list double
           elif isinstance(value, list):
               for el in value:
                 if not isinstance(el, list):
                   return False
                 for e in el:
-                  if not isinstance(e, float):
+                  if not isinstance(e, double):
                     return False
               return True
           return False
